Refactor GetInterestPerBlockHighPrecisionString
