+++ src/masternodes/rpc_loan.cpp
@@ -1,5 +1,4 @@
 #include <masternodes/mn_rpc.h>
-#include <boost/multiprecision/cpp_int.hpp>
 
 extern UniValue tokenToJSON(DCT_ID const& id, CTokenImplementation const& token, bool verbose);
 extern UniValue listauctions(const JSONRPCRequest& request);
@@ -1324,44 +1323,6 @@ UniValue getloaninfo(const JSONRPCRequest& request) {
     ret.pushKV("totals", totalsObj);
 
     return (ret);
-}
-
-std::string GetInterestPerBlockHighPrecisionString(base_uint<128> value) {
-    struct HighPrecisionInterestValue {
-        typedef boost::multiprecision::int128_t int128;
-        typedef int64_t int64;
-
-        int128 value;
-
-        HighPrecisionInterestValue(base_uint<128> val) {
-            value = int128("0x" + val.GetHex());
-        }
-
-        int64 GetInterestPerBlockSat() {
-            return int64(value / HIGH_PRECISION_SCALER);
-        }
-
-        int64 GetInterestPerBlockSubSat() {
-            return int64(value % HIGH_PRECISION_SCALER);
-        }
-
-        int64 GetInterestPerBlockMagnitude() {
-            return int64(value / HIGH_PRECISION_SCALER / COIN);
-        }
-
-        int128 GetInterestPerBlockDecimal() {
-            auto v = GetInterestPerBlockSat();
-            return v == 0 ? value : value % (int128(HIGH_PRECISION_SCALER) * COIN);
-        }
-
-        std::string GetInterestPerBlockString() {
-            std::ostringstream result;
-            result << GetInterestPerBlockMagnitude() << ".";
-            result << std::setw(24) << std::setfill('0') << GetInterestPerBlockDecimal();
-            return result.str();
-        }
-    };
-    return HighPrecisionInterestValue(value).GetInterestPerBlockString();
 }
 
 UniValue getinterest(const JSONRPCRequest& request) {
